requirements.txt
# Pinned for reproducible builds (Python 3.11)
fastapi==0.115.6
uvicorn[standard]==0.32.1
langchain==0.3.13
langchain-core==0.3.28
langgraph==0.2.60
langchain-openai==0.2.14
openai==1.59.6
pydantic==2.10.4
pydantic-settings==2.6.1
boto3==1.35.86
botocore==1.35.86
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
watchtower==3.3.1
